GDM/requirements.txt
# GitHub & GitHub Enterprise Repository Fetcher - Requirements
# ==============================================================

# Core dependencies
pydantic>=1.10.0,<3.0.0
pyautogen>=0.2.0

# Web framework and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Optional: Streamlit interface (alternative)
streamlit>=1.28.0

# HTTP and networking
requests>=2.25.0
urllib3>=1.26.0
httpx>=0.24.0

# GitHub API and Authentication
PyGithub>=1.59.0
python-jose[cryptography]>=3.3.0
cryptography>=3.4.8
python-multipart>=0.0.6

# CLI enhancements
rich>=13.0.0
click>=8.0.0
typer>=0.9.0

# Database (for token storage)
sqlalchemy>=1.4.0
aiosqlite>=0.17.0

# Environment management
python-dotenv>=0.19.0

# Development dependencies (optional)
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0
# pytest>=7.0.0
# pytest-asyncio>=0.21.0